# angular-src/src/app/swagger-api/model/iErrorResponse.ts
/**
 * TSOA-Node Express
 * TSOA-NodeExpress Documentation
 *
 * OpenAPI spec version: 1.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { IMongoError } from './iMongoError';


export interface IErrorResponse {
    thrown?: boolean;
    error?: IMongoError;
    message?: string;
}