# angular-src/src/app/swagger-api/model/iLoginParams.ts
/**
 * TSOA-Node Express
 * TSOA-NodeExpress Documentation
 *
 * OpenAPI spec version: 1.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ILoginParams {
    username?: string;
    email?: string;
    password?: string;
}