# angular-src/src/app/swagger-api/model/iTaskResponse.ts
/**
 * TSOA-Node Express
 * TSOA-NodeExpress Documentation
 *
 * OpenAPI spec version: 1.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { IUserVm } from './iUserVm';


export interface ITaskResponse {
    _id?: string;
    title?: string;
    content?: string;
    slug?: string;
    createdOn?: Date;
    updatedOn?: Date;
    isCompleted?: boolean;
    user?: IUserVm;
}