# angular-src/src/app/swagger-api/model/iTaskVm.ts
/**
 * TSOA-Node Express
 * TSOA-NodeExpress Documentation
 *
 * OpenAPI spec version: 1.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ITaskVm {
    _id?: string;
    title?: string;
    slug?: string;
    content?: string;
    createdOn?: Date;
    updatedOn?: Date;
    isCompleted?: boolean;
}