# angular-src/src/app/swagger-api/model/iUserVm.ts
/**
 * TSOA-Node Express
 * TSOA-NodeExpress Documentation
 *
 * OpenAPI spec version: 1.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { IUserProfile } from './iUserProfile';
import { UserRole } from './userRole';


export interface IUserVm {
    _id?: string;
    username: string;
    email: string;
    password?: string;
    createdOn?: Date;
    updatedOn?: Date;
    role?: UserRole;
    lastVisited?: Date;
    profile?: IUserProfile;
}